Move process_events_iter to picard.util.qt
process_events_iter drives the Qt event loop via QCoreApplication, so it
moves to the Qt-only picard.util.qt. Update the importer (tagger).
