+++ picard/util/qt.py
@@ -26,6 +26,12 @@ keeps the bulk of the utility helpers usable from Qt-free contexts (e.g. build
 tooling or a potential headless/alternative frontend).
 """
 
+from collections.abc import (
+    Iterable,
+    Iterator,
+)
+from time import monotonic
+
 from PyQt6 import QtCore
 
 
@@ -59,3 +65,27 @@ class ReadWriteLockContext:
 
     def __exit__(self, type, value, tb):
         self.__lock.unlock()
+
+
+def process_events_iter(iterable: Iterable, interval: float = 0.1) -> Iterator:
+    """
+    Creates an iterator over iterable that calls QCoreApplication.processEvents()
+    after certain time intervals.
+
+    This must only be used in the main thread.
+
+    Args:
+        iterable: iterable object to iterate over
+        interval: interval in seconds to call QCoreApplication.processEvents()
+    """
+    if interval:
+        start = monotonic()
+    for item in iterable:
+        if interval:
+            now = monotonic()
+            delta = now - start
+            if delta > interval:
+                start = now
+                QtCore.QCoreApplication.processEvents()
+        yield item
+    QtCore.QCoreApplication.processEvents()
